move exceptions to a separate submodule
