+++ autointent/modules/abc/__init__.py
@@ -1,6 +1,6 @@
-from ._base import Module, WrongClassificationError
+from ._base import Module
 from ._decision import DecisionModule
 from ._embedding import EmbeddingModule
 from ._scoring import ScoringModule
 
-__all__ = ["DecisionModule", "EmbeddingModule", "Module", "ScoringModule", "WrongClassificationError"]
+__all__ = ["DecisionModule", "EmbeddingModule", "Module", "ScoringModule"]
